fix default_value type error

+++ ConfigSpace/hyperparameters/uniform_float.pyx
@@ -118,13 +118,13 @@ cdef class UniformFloatHyperparameter(FloatHyperparameter):
     def check_default(self, default_value: Optional[float]) -> float:
         if default_value is None:
             if self.log:
-                default_value = np.exp((np.log(self.lower) + np.log(self.upper)) / 2.)
+                default_value = float(np.exp((np.log(self.lower) + np.log(self.upper)) / 2.))
             else:
-                default_value = (self.lower + self.upper) / 2.
-        default_value = np.round(float(default_value), 10)
+                default_value = float((self.lower + self.upper) / 2.)
+        default_value = float(np.round(float(default_value), 10))
 
         if self.is_legal(default_value):
-            return default_value
+            return float(default_value)
         else:
             raise ValueError("Illegal default value %s" % str(default_value))
